+++ tests/helpers/host-component.mjs
@@ -1,0 +1,7 @@
+import { createElement } from "react";
+
+export function hostComponent(type) {
+  return function HostComponent(props) {
+    return createElement(type, props, props.children);
+  };
+}